Make commit messages great again!
Move more deprecated tile checking code to the tile converter class
